+++ public/js/auth-page.js
@@ -3,11 +3,12 @@ const fullName = document.querySelector('#name')
 const username = document.querySelector('#username')
 const password = document.querySelector('#password')
 const email = document.querySelector('#email')
-// const display = document.querySelector('#error')
+const display = document.querySelector('#error')
 const submitBtn = document.querySelector('#submit').value.toLowerCase()
+
 form.addEventListener('submit', async (e) => {
     e.preventDefault()
-    // display.textContent = ''
+    display.textContent = ''
 
     // if Login Page
     if (submitBtn == 'login') {
@@ -17,14 +18,14 @@ form.addEventListener('submit', async (e) => {
         });
     }
 
-
     // If Register user page 
     if (submitBtn == 'register') {
         authApi('register', bodyOptions = {
             name:  fullName.value,
             username: username.value,
             password: password.value,
-            email: email.value
+            email: email.value,
+            role: 'basic'
         });
     }
 })
@@ -39,12 +40,14 @@ async function authApi(apiEndPoint, bodyOptions) {
         })
         const data = await res.json()
         if (res.status === 400 || res.status === 401) {
-            // display.classList.add("error")
-            // console.log(display)
-            // return display.textContent = `${data.error.code}. ${(data.error.code != 100) ? data.error.message : ''}`
-            return alert(`${data.error.code}. ${(data.error.code != 100) ? data.error.message : ''}`)
+            display.classList.add("error")
+            display.textContent = `Error ${data.error.code} - ${(data.error.code != 100) ? data.error.message : ''}`
+            setTimeout(() => {
+                display.classList.remove('error')
+            }, 10 * 1000)
+            return;
         }
-        data.role === "basic" ? window.location.replace(document.referrer) : window.location.replace(document.referrer)
+        data.role === "basic" ? history.back() : history.back()
     } catch (err) {
         console.log(err.message)
     }